Add files via upload
This project here is pretty much almost done. I was able to scrape data from the website : https://wiki.serenesforest.net/index.php/Fire_Emblem_TCG. I was able to create a Flask application that had cleaned some issues in the data such as the cards' id does not match with the actual cards or card name has issues with duplicates or having different strings like (Male) or (Judgral).  All you have to do to run the file is go to Project_1/P1_MAIN/app.py and go to terminal and run the command py app.py. It is that simple just run : 

To run in command prompt : 
1.  cd Project_1
2. cd P1_MAIN
3. py app.py 
This is how you will be able to see the dashboard where user gets to see things like the different rares for a card and a histogram. 

Upcoming features will basically do final cleaning like having only in 1 excel file instead of 2 that way when showing the dashboard or parsing the data , I would only have to use 2 dataframes instead of 1 .

Hopefully this can be done by May.